MainMenu QA'd
QA of the main menu:
Buttons:
a). Buttons should be a solid colour
b). Buttons should change to a different colour when mouse is hovered over it
c). Buttons should change to a different colour when selected using left mouse
click.
d). When a button is selected an appropriate debug messages should appear in the console e.g. (Play Button Pressed!)
Implement buttons for Play, Settings, and Instructions using the above criteria.