Feeding Icon Always Visible
Fixed the issue where the feeding icon would sometimes not appear when you were feeding. This was a result of pausing the game and going back to the main menu, then going back in-game which would cause the health bar reference to be lost, therefore the feeding animation didn't play anymore.